Refactoring reducers
