+++ app/frontend/actions/openEditMapDialog.js
@@ -1,8 +1,11 @@
 import { OPEN_EDIT_MAP_DIALOG } from '../actionTypes';
 
-const openEditMapDialog = () => {
+const openEditMapDialog = (map) => {
   return {
-    type: OPEN_EDIT_MAP_DIALOG
+    type: OPEN_EDIT_MAP_DIALOG,
+    payload: {
+      map: map
+    }
   }
 }
